Adding config to topmost layer
The builder seemed to drop the keys if they were not in this top most
layer, i suppose this is due to the merge. So simply forward port them
into the topmost beat layer.

Refactors reactive/filebeat.py to not rely on the older states of
beat.repo.available and filebeat.installed